added LibAccount LibNotif and AccountMapper

+++ Client-bank/src/main/java/com/mycompany/client/bank/jpa/Account.java
@@ -66,7 +66,7 @@ public class Account implements Serializable {
 		this.accountId = accountId;
 	}
 
-	public Account(Long accountId, int value, Date openDate) {
+	public Account(Long accountId, double value, Date openDate) {
 		this.accountId = accountId;
 		this.value = value;
 		this.openDate = openDate;
